H001 Sell in partialities As Owner, I want to create sales that can be payed in partialities until it's completly payed To be able to sell expensive items with no credit card.
H001.03 Partial payments
We need to accept adicional payments to a DUED sale. This way, our client are able to pay off their debt.

To acomplish this, we need to retrieve the DUED sale and to Sales Register the same way Unsuspend button does.
The tricky part here is that it also retrieve previous payments and it must be deleted to add new ones. So, this payments rules must change in order to show previous payments without due ones and they have to accept two same method payments

Payment type was modified to concat the current date of payment. This way, is possible to add multiple payments